flask/requirements.txt
flask>=2.3.0
requests>=2.31.0
PyPDF2==3.0.1
docx2txt==0.8
sentence-transformers==3.0.1
torch>=2.0.0
pandas>=1.5.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
